+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.5] - 2024-03-17
+
 ### Added
  - Implement `TryFrom<&str>` for `{Array,Group}Metadata`
 
@@ -685,7 +687,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Initial public release
 
-[unreleased]: https://github.com/LDeakin/zarrs/compare/v0.12.4...HEAD
+[unreleased]: https://github.com/LDeakin/zarrs/compare/v0.12.5...HEAD
+[0.12.5]: https://github.com/LDeakin/zarrs/releases/tag/v0.12.5
 [0.12.4]: https://github.com/LDeakin/zarrs/releases/tag/v0.12.4
 [0.12.3]: https://github.com/LDeakin/zarrs/releases/tag/v0.12.3
 [0.12.2]: https://github.com/LDeakin/zarrs/releases/tag/v0.12.2